Prefer the discounted price over standard if the both exist
This and the previous way have the unfortunate side affect of always preferring and displaying one price style over another, even if this is incorrect. A discounted price will be shown over the normal price, even if the normal price is the real one. Look into fixing this if we find a good steam API for it.